refactor: move Color to fieldofview.py
